Refatoração - WIP
