+++ src/modules/controle_aula/index.js
@@ -4,13 +4,6 @@ import * as formAddAulas from "./formAddAulas.js";
 import * as formAddCursos from "./formAddCursos.js";
 import * as formAddAlunos from "./formAddAlunos.js";
 
-//========================================================================================================
-//======================================= FORM ===========================================================
-//========================================================================================================
-
-
-
-//--------------------Carrega funções----------------------------
 //TODO: Arrumar ordem de execução das funções
 function loadOnStartUp(){
   insertAulasCursosFunc.insertAulasWhenChangeAluno();
@@ -24,19 +17,17 @@ insertAulasCursosFunc.insertAulasWhenAlunoChange
   formAddAulas.eventFormsAdd();
   formAddAulas.AddEventBtnCloseForm();
   formAddAulas.eventSelectAlunoAddAula();
-  
+
   //Carrega a lista de cursos do primeiro aluno quando inciado
   //no formulário form_add_aulas
   formAddAulas.insertSelectCursosAddAula("RA01");
   
   //Curso
   formAddCursos.eventSelectAlunoAddCurso();
-  //TODO: Conferir utilidade dessa chamada de função
   
   //Aluno
-  formAddAlunos.validaSelectOptionsAddAluno();
   formAddAlunos.insertOptionsAddAlunoRA();
-
+  formAddAlunos.validaSelectOptionsAddAluno();
 
 }
 loadOnStartUp()
